requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
mysqlclient==1.4.6
Flask-MySQLdb==0.2.0
Marshmallow==3.5.2
SQLAlchemy==1.3.16
flask-marshmallow==0.12.0
flask-sqlalchemy==2.4.1
Flask-Executor==0.9.3
marshmallow-sqlalchemy==0.23.0
pyJWT==1.7.1
pytest==5.4.2
Pillow==7.0.0
opencv-contrib-python==4.2.0.34
